Cleanup duplicate logic and refactor installations.js into InstallationsRouter.
